requirements.txt
mcp>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
neo4j>=5.0.0
pydantic>=2.0
tiktoken>=0.5.0
pyyaml>=6.0
watchdog>=3.0.0